refactor: remove namespace, remove needless exports, add some jsdoc
