+++ components/clarinet-deno/index.ts
@@ -1,5 +1,4 @@
-// deno-lint-ignore-file ban-ts-comment no-namespace
-
+// deno-lint-ignore-file ban-ts-comment
 import {
   ExpectFungibleTokenBurnEvent,
   ExpectFungibleTokenMintEvent,
@@ -10,9 +9,11 @@ import {
   ExpectPrintEvent,
   ExpectSTXTransferEvent,
   ExpectSTXBurnEvent,
-} from "./types.ts";
+} from "./eventTypes.ts";
+import * as types from "./clarityTypes.ts";
 
-export * from "./types.ts";
+export * from "./eventTypes.ts";
+export * as types from "./clarityTypes.ts";
 
 export class Tx {
   type: number;
@@ -133,11 +134,10 @@ export class Chain {
       })
     );
     this.blockHeight = result.block_height;
-    const block: Block = {
+    return {
       height: result.block_height,
       receipts: result.receipts,
     };
-    return block;
   }
 
   mineEmptyBlock(count: number): EmptyBlock {
@@ -145,15 +145,14 @@ export class Chain {
       // @ts-ignore
       Deno.core.opSync("api/v1/mine_empty_blocks", {
         sessionId: this.sessionId,
-        count: count,
+        count,
       })
     );
     this.blockHeight = result.block_height;
-    const emptyBlock: EmptyBlock = {
+    return {
       session_id: result.session_id,
       block_height: result.block_height,
     };
-    return emptyBlock;
   }
 
   mineEmptyBlockUntil(targetBlockHeight: number): EmptyBlock {
@@ -166,28 +165,35 @@ export class Chain {
     return this.mineEmptyBlock(count);
   }
 
+  /**
+   * Call a read-only function
+   * @param contract Address of the contract implementing the function
+   * @param method The read-only function to call
+   * @param args Arguments to pass as clarity values
+   * @param sender Address of the caller
+   * @returns The result of th
+   */
   callReadOnlyFn(
     contract: string,
     method: string,
-    args: Array<unknown>,
+    args: Array<string>,
     sender: string
   ): ReadOnlyFn {
     const result = JSON.parse(
       // @ts-ignore
       Deno.core.opSync("api/v1/call_read_only_fn", {
         sessionId: this.sessionId,
-        contract: contract,
-        method: method,
-        args: args,
-        sender: sender,
+        contract,
+        method,
+        args,
+        sender,
       })
     );
-    const readOnlyFn: ReadOnlyFn = {
+    return {
       session_id: result.session_id,
       result: result.result,
       events: result.events,
     };
-    return readOnlyFn;
   }
 
   getAssetsMaps(): AssetsMaps {
@@ -197,11 +203,10 @@ export class Chain {
         sessionId: this.sessionId,
       })
     );
-    const assetsMaps: AssetsMaps = {
+    return {
       session_id: result.session_id,
       assets: result.assets,
     };
-    return assetsMaps;
   }
 
   switchEpoch(epoch: string): boolean {
@@ -209,7 +214,7 @@ export class Chain {
       // @ts-ignore
       Deno.core.opSync("api/v1/switch_epoch", {
         sessionId: this.sessionId,
-        epoch: epoch,
+        epoch,
       })
     );
     return result;
@@ -236,10 +241,21 @@ interface UnitTestOptions {
   fn: TestFunction;
 }
 
+interface FunctionInterface {
+  name: string;
+  access: "read_only" | "public" | "private";
+  args: {
+    name: string;
+    type: string;
+  }[];
+}
+
 export interface Contract {
   contract_id: string;
   source: string;
-  contract_interface: unknown;
+  contract_interface: {
+    functions: FunctionInterface[];
+  };
 }
 
 export interface StacksNode {
@@ -340,93 +356,6 @@ export class Clarinet {
         await options.fn(accounts, contracts, stacks_node);
       },
     });
-  }
-}
-
-export namespace types {
-  const byteToHex: string[] = [];
-  for (let n = 0; n <= 0xff; ++n) {
-    const hexOctet = n.toString(16).padStart(2, "0");
-    byteToHex.push(hexOctet);
-  }
-
-  function serializeTuple(input: Record<string, unknown>) {
-    const items: Array<string> = [];
-    for (const [key, value] of Object.entries(input)) {
-      if (Array.isArray(value)) {
-        throw new Error("Tuple value can't be an array");
-      } else if (!!value && typeof value === "object") {
-        items.push(
-          `${key}: { ${serializeTuple(value as Record<string, unknown>)} }`
-        );
-      } else {
-        items.push(`${key}: ${value}`);
-      }
-    }
-    return items.join(", ");
-  }
-
-  export function ok(val: string) {
-    return `(ok ${val})`;
-  }
-
-  export function err(val: string) {
-    return `(err ${val})`;
-  }
-
-  export function some(val: string) {
-    return `(some ${val})`;
-  }
-
-  export function none() {
-    return `none`;
-  }
-
-  export function bool(val: boolean) {
-    return `${val}`;
-  }
-
-  export function int(val: number | bigint) {
-    return `${val}`;
-  }
-
-  export function uint(val: number | bigint) {
-    return `u${val}`;
-  }
-
-  export function ascii(val: string) {
-    return JSON.stringify(val);
-  }
-
-  export function utf8(val: string) {
-    return `u${JSON.stringify(val)}`;
-  }
-
-  export function buff(val: ArrayBuffer | string) {
-    const buff =
-      typeof val == "string"
-        ? new TextEncoder().encode(val)
-        : new Uint8Array(val);
-
-    const hexOctets = new Array(buff.length);
-
-    for (let i = 0; i < buff.length; ++i) {
-      hexOctets[i] = byteToHex[buff[i]];
-    }
-
-    return `0x${hexOctets.join("")}`;
-  }
-
-  export function list(val: Array<unknown>) {
-    return `(list ${val.join(" ")})`;
-  }
-
-  export function principal(val: string) {
-    return `'${val}`;
-  }
-
-  export function tuple(val: Record<string, unknown>) {
-    return `{ ${serializeTuple(val)} }`;
   }
 }
 
@@ -921,10 +850,10 @@ function run(str: string, code: Code): string {
     : str;
 }
 
-export function red(str: string): string {
+function red(str: string): string {
   return run(str, code([31], 39));
 }
 
-export function green(str: string): string {
+function green(str: string): string {
   return run(str, code([32], 39));
 }
